requirements.txt
requests
pytest
ruff
python-dotenv
